pyproject.toml
[tool.ruff]
line-length = 100
target-version = "py311"  # Match common WSL2 Ubuntu Python version

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

[tool.ruff.lint.isort]
known-first-party = ["datakit_bronze", "datakit_postgres", "datakit_sqlserver", "datakit_spark"]
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# Project metadata (for future Python packaging if needed)
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "workstation-setup"
version = "1.0.0"
description = "Comprehensive workstation setup for Astronomer Airflow data engineering platform"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Data Engineering Team"},
]
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["wizard*", "platform_lib"]
exclude = ["kerberos*", "openmetadata*", "deprecated*", "tests*"]

[tool.pytest.ini_options]
testpaths = ["wizard"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
# Don't collect from implementation modules, only from test modules
norecursedirs = [".*", "build", "dist", "__pycache__"]
